4 commit

+++ src/com/company/Division.java
@@ -3,6 +3,6 @@ package com.company;
 public class Division implements Operation {
     @Override
     public double calculate(int a, int b) {
-        return (double) a/(double) b;
+        return (double) a/ b;
     }
 }
